# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Photography Studio API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type PackageCategory =
  (typeof PackageCategory)[keyof typeof PackageCategory];

export const PackageCategory = {
  portrait: "portrait",
  wedding: "wedding",
  event: "event",
  commercial: "commercial",
  family: "family",
} as const;

export interface Package {
  id: number;
  name: string;
  description: string;
  price: number;
  duration: string;
  features: string[];
  category: PackageCategory;
  popular: boolean;
  createdAt: string;
}

export type BookingStatus = (typeof BookingStatus)[keyof typeof BookingStatus];

export const BookingStatus = {
  pending: "pending",
  confirmed: "confirmed",
  completed: "completed",
  cancelled: "cancelled",
} as const;

export interface Booking {
  id: number;
  customerName: string;
  customerEmail: string;
  customerPhone: string;
  packageId: number;
  packageName: string;
  packagePrice: number;
  eventDate: string;
  eventType: string;
  location: string;
  notes?: string;
  status: BookingStatus;
  invoiceNumber: string;
  createdAt: string;
}

export interface CreateBookingRequest {
  /** @minLength 2 */
  customerName: string;
  customerEmail: string;
  /** @minLength 7 */
  customerPhone: string;
  packageId: number;
  eventDate: string;
  eventType: string;
  location: string;
  notes?: string;
}

export type UpdateBookingStatusRequestStatus =
  (typeof UpdateBookingStatusRequestStatus)[keyof typeof UpdateBookingStatusRequestStatus];

export const UpdateBookingStatusRequestStatus = {
  pending: "pending",
  confirmed: "confirmed",
  completed: "completed",
  cancelled: "cancelled",
} as const;

export interface UpdateBookingStatusRequest {
  status: UpdateBookingStatusRequestStatus;
}

export interface ContactRequest {
  /** @minLength 2 */
  name: string;
  email: string;
  phone?: string;
  subject: string;
  /** @minLength 10 */
  message: string;
}

export interface ContactMessage {
  id: number;
  name: string;
  email: string;
  phone?: string;
  subject: string;
  message: string;
  read: boolean;
  createdAt: string;
}

export interface Testimonial {
  id: number;
  clientName: string;
  clientTitle: string;
  /**
   * @minimum 1
   * @maximum 5
   */
  rating: number;
  content: string;
  published: boolean;
  createdAt: string;
}

export interface CreateTestimonialRequest {
  clientName: string;
  clientTitle: string;
  /**
   * @minimum 1
   * @maximum 5
   */
  rating: number;
  content: string;
  published?: boolean;
}

export interface AdminStats {
  totalBookings: number;
  pendingBookings: number;
  confirmedBookings: number;
  completedBookings: number;
  totalRevenue: number;
  unreadMessages: number;
  totalTestimonials: number;
}

export interface ErrorResponse {
  error: string;
  details?: string[];
}