# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Photography Studio API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /bookings:
    get:
      operationId: getBookings
      tags: [bookings]
      summary: Get all bookings (admin)
      responses:
        "200":
          description: List of bookings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Booking"
    post:
      operationId: createBooking
      tags: [bookings]
      summary: Create a new booking
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBookingRequest"
      responses:
        "201":
          description: Booking created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /bookings/{id}:
    get:
      operationId: getBookingById
      tags: [bookings]
      summary: Get a booking by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Booking found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateBookingStatus
      tags: [bookings]
      summary: Update booking status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBookingStatusRequest"
      responses:
        "200":
          description: Updated booking
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
...
components:
  schemas:
...
    Booking:
      type: object
      properties:
        id:
          type: integer
        customerName:
          type: string
        customerEmail:
          type: string
        customerPhone:
          type: string
        packageId:
          type: integer
        packageName:
          type: string
        packagePrice:
          type: number
        eventDate:
          type: string
          format: date
        eventType:
          type: string
        location:
          type: string
        notes:
          type: string
        status:
          type: string
          enum: [pending, confirmed, completed, cancelled]
        invoiceNumber:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - customerName
        - customerEmail
        - customerPhone
        - packageId
        - packageName
        - packagePrice
        - eventDate
        - eventType
        - location
        - status
        - invoiceNumber
        - createdAt
    CreateBookingRequest:
      type: object
      properties:
        customerName:
          type: string
          minLength: 2
        customerEmail:
          type: string
          format: email
        customerPhone:
          type: string
          minLength: 7
        packageId:
          type: integer
        eventDate:
          type: string
          format: date
        eventType:
          type: string
        location:
          type: string
        notes:
          type: string
      required:
        - customerName
        - customerEmail
        - customerPhone
        - packageId
        - eventDate
        - eventType
        - location
    UpdateBookingStatusRequest:
      type: object
      properties:
        status:
          type: string
          enum: [pending, confirmed, completed, cancelled]
      required:
        - status
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        details:
          type: array
          items:
            type: string
      required:
        - error