# lib/api-zod/src/generated/types/booking.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Photography Studio API
 * OpenAPI spec version: 0.1.0
 */
import type { BookingStatus } from "./bookingStatus";

export interface Booking {
  id: number;
  customerName: string;
  customerEmail: string;
  customerPhone: string;
  packageId: number;
  packageName: string;
  packagePrice: number;
  eventDate: Date;
  eventType: string;
  location: string;
  notes?: string;
  status: BookingStatus;
  invoiceNumber: string;
  createdAt: Date;
}