# lib/api-zod/src/generated/types/createBookingRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Photography Studio API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateBookingRequest {
  /** @minLength 2 */
  customerName: string;
  customerEmail: string;
  /** @minLength 7 */
  customerPhone: string;
  packageId: number;
  eventDate: Date;
  eventType: string;
  location: string;
  notes?: string;
}